+++ cfg/yolo.cfg
@@ -9,8 +9,8 @@ decay=0.0005
 
 learning_rate=0.001
 policy=steps
-steps=100,200,300,400,500,600,700,20000,30000
-scales=2,2,1.25,1.25,1.25,1.25,1.03,.1,.1
+steps=200,400,600,20000,30000
+scales=2.5,2,2,.1,.1
 max_batches = 40000
 
 [crop]
@@ -218,7 +218,7 @@ probability=.5
 output= 1470
 activation=linear
 
-[region]
+[detection]
 classes=20
 coords=4
 rescore=1
