+++ src/detection_layer.c
@@ -6,42 +6,32 @@
 #include "cuda.h"
 #include "utils.h"
 #include <stdio.h>
+#include <assert.h>
 #include <string.h>
 #include <stdlib.h>
 
-int get_detection_layer_locations(detection_layer l)
-{
-    return l.inputs / (l.classes+l.coords+l.joint+(l.background || l.objectness));
-}
-
-int get_detection_layer_output_size(detection_layer l)
-{
-    return get_detection_layer_locations(l)*((l.background || l.objectness) + l.classes + l.coords);
-}
-
-detection_layer make_detection_layer(int batch, int inputs, int classes, int coords, int joint, int rescore, int background, int objectness)
+detection_layer make_detection_layer(int batch, int inputs, int n, int side, int classes, int coords, int rescore)
 {
     detection_layer l = {0};
     l.type = DETECTION;
-    
+
+    l.n = n;
     l.batch = batch;
     l.inputs = inputs;
     l.classes = classes;
     l.coords = coords;
     l.rescore = rescore;
-    l.objectness = objectness;
-    l.background = background;
-    l.joint = joint;
+    l.side = side;
+    assert(side*side*((1 + l.coords)*l.n + l.classes) == inputs);
     l.cost = calloc(1, sizeof(float));
-    l.does_cost=1;
-    int outputs = get_detection_layer_output_size(l);
-    l.outputs = outputs;
-    l.output = calloc(batch*outputs, sizeof(float));
-    l.delta = calloc(batch*outputs, sizeof(float));
-    #ifdef GPU
-    l.output_gpu = cuda_make_array(l.output, batch*outputs);
-    l.delta_gpu  = cuda_make_array(l.delta,  batch*outputs);
-    #endif
+    l.outputs = l.inputs;
+    l.truths = l.side*l.side*(1+l.coords+l.classes);
+    l.output = calloc(batch*l.outputs, sizeof(float));
+    l.delta = calloc(batch*l.outputs, sizeof(float));
+#ifdef GPU
+    l.output_gpu = cuda_make_array(l.output, batch*l.outputs);
+    l.delta_gpu = cuda_make_array(l.delta, batch*l.outputs);
+#endif
 
     fprintf(stderr, "Detection Layer\n");
     srand(0);
@@ -51,124 +41,164 @@ detection_layer make_detection_layer(int batch, int inputs, int classes, int coo
 
 void forward_detection_layer(const detection_layer l, network_state state)
 {
-    int in_i = 0;
-    int out_i = 0;
-    int locations = get_detection_layer_locations(l);
+    int locations = l.side*l.side;
     int i,j;
-    for(i = 0; i < l.batch*locations; ++i){
-        int mask = (!state.truth || state.truth[out_i + (l.background || l.objectness) + l.classes + 2]);
-        float scale = 1;
-        if(l.joint) scale = state.input[in_i++];
-        else if(l.objectness){
-            l.output[out_i++] = 1-state.input[in_i++];
-            scale = mask;
-        }
-        else if(l.background) l.output[out_i++] = scale*state.input[in_i++];
-
-        for(j = 0; j < l.classes; ++j){
-            l.output[out_i++] = scale*state.input[in_i++];
-        }
-        if(l.objectness){
-
-        }else if(l.background){
-            softmax_array(l.output + out_i - l.classes-l.background, l.classes+l.background, l.output + out_i - l.classes-l.background);
-            activate_array(state.input+in_i, l.coords, LOGISTIC);
-        }
-        for(j = 0; j < l.coords; ++j){
-            l.output[out_i++] = mask*state.input[in_i++];
+    memcpy(l.output, state.input, l.outputs*l.batch*sizeof(float));
+    int b;
+    if (l.softmax){
+        for(b = 0; b < l.batch; ++b){
+            int index = b*l.inputs;
+            for (i = 0; i < locations; ++i) {
+                int offset = i*l.classes;
+                softmax_array(l.output + index + offset, l.classes,
+                        l.output + index + offset);
+            }
+            int offset = locations*l.classes;
+            activate_array(l.output + index + offset, locations*l.n*(1+l.coords), LOGISTIC);
         }
     }
-    float avg_iou = 0;
-    int count = 0;
-    if(l.does_cost && state.train){
+    if(state.train){
+        float avg_iou = 0;
+        float avg_cat = 0;
+        float avg_allcat = 0;
+        float avg_obj = 0;
+        float avg_anyobj = 0;
+        int count = 0;
         *(l.cost) = 0;
-        int size = get_detection_layer_output_size(l) * l.batch;
+        int size = l.inputs * l.batch;
         memset(l.delta, 0, size * sizeof(float));
-        for (i = 0; i < l.batch*locations; ++i) {
-            int classes = (l.objectness || l.background)+l.classes;
-            int offset = i*(classes+l.coords);
-            for (j = offset; j < offset+classes; ++j) {
-                *(l.cost) += pow(state.truth[j] - l.output[j], 2);
-                l.delta[j] =  state.truth[j] - l.output[j];
-                if(l.background && j == offset) l.delta[j] *= .1;
-            }
-
-            box truth;
-            truth.x = state.truth[j+0]/7;
-            truth.y = state.truth[j+1]/7;
-            truth.w = pow(state.truth[j+2], 2);
-            truth.h = pow(state.truth[j+3], 2);
-
-            box out;
-            out.x = l.output[j+0]/7;
-            out.y = l.output[j+1]/7;
-            out.w = pow(l.output[j+2], 2);
-            out.h = pow(l.output[j+3], 2);
-
-            if(!(truth.w*truth.h)) continue;
-            float iou = box_iou(out, truth);
-            avg_iou += iou;
-            ++count;
-
-            *(l.cost) += pow((1-iou), 2);
-            l.delta[j+0] = 4 * (state.truth[j+0] - l.output[j+0]);
-            l.delta[j+1] = 4 * (state.truth[j+1] - l.output[j+1]);
-            l.delta[j+2] = 4 * (state.truth[j+2] - l.output[j+2]);
-            l.delta[j+3] = 4 * (state.truth[j+3] - l.output[j+3]);
-            if(l.rescore){
-                if(l.objectness){
-                    state.truth[offset] = iou;
-                    l.delta[offset] = state.truth[offset] - l.output[offset];
+        for (b = 0; b < l.batch; ++b){
+            int index = b*l.inputs;
+            for (i = 0; i < locations; ++i) {
+                int truth_index = (b*locations + i)*(1+l.coords+l.classes);
+                int is_obj = state.truth[truth_index];
+                for (j = 0; j < l.n; ++j) {
+                    int p_index = index + locations*l.classes + i*l.n + j;
+                    l.delta[p_index] = l.noobject_scale*(0 - l.output[p_index]);
+                    *(l.cost) += l.noobject_scale*pow(l.output[p_index], 2);
+                    avg_anyobj += l.output[p_index];
                 }
-                else{
-                    for (j = offset; j < offset+classes; ++j) {
-                        if(state.truth[j]) state.truth[j] = iou;
-                        l.delta[j] =  state.truth[j] - l.output[j];
+
+                int best_index = -1;
+                float best_iou = 0;
+                float best_rmse = 20;
+
+                if (!is_obj){
+                    continue;
+                }
+
+                int class_index = index + i*l.classes;
+                for(j = 0; j < l.classes; ++j) {
+                    l.delta[class_index+j] = l.class_scale * (state.truth[truth_index+1+j] - l.output[class_index+j]);
+                    *(l.cost) += l.class_scale * pow(state.truth[truth_index+1+j] - l.output[class_index+j], 2);
+                    if(state.truth[truth_index + 1 + j]) avg_cat += l.output[class_index+j];
+                    avg_allcat += l.output[class_index+j];
+                }
+
+                box truth = float_to_box(state.truth + truth_index + 1 + l.classes);
+                truth.x /= l.side;
+                truth.y /= l.side;
+
+                for(j = 0; j < l.n; ++j){
+                    int box_index = index + locations*(l.classes + l.n) + (i*l.n + j) * l.coords;
+                    box out = float_to_box(l.output + box_index);
+                    out.x /= l.side;
+                    out.y /= l.side;
+
+                    if (l.sqrt){
+                        out.w = out.w*out.w;
+                        out.h = out.h*out.h;
+                    }
+
+                    float iou  = box_iou(out, truth);
+                    //iou = 0;
+                    float rmse = box_rmse(out, truth);
+                    if(best_iou > 0 || iou > 0){
+                        if(iou > best_iou){
+                            best_iou = iou;
+                            best_index = j;
+                        }
+                    }else{
+                        if(rmse < best_rmse){
+                            best_rmse = rmse;
+                            best_index = j;
+                        }
                     }
                 }
+
+                if(l.forced){
+                    if(truth.w*truth.h < .1){
+                        best_index = 1;
+                    }else{
+                        best_index = 0;
+                    }
+                }
+
+                int box_index = index + locations*(l.classes + l.n) + (i*l.n + best_index) * l.coords;
+                int tbox_index = truth_index + 1 + l.classes;
+
+                box out = float_to_box(l.output + box_index);
+                out.x /= l.side;
+                out.y /= l.side;
+                if (l.sqrt) {
+                    out.w = out.w*out.w;
+                    out.h = out.h*out.h;
+                }
+                float iou  = box_iou(out, truth);
+
+                //printf("%d", best_index);
+                int p_index = index + locations*l.classes + i*l.n + best_index;
+                *(l.cost) -= l.noobject_scale * pow(l.output[p_index], 2);
+                *(l.cost) += l.object_scale * pow(1-l.output[p_index], 2);
+                avg_obj += l.output[p_index];
+                l.delta[p_index] = l.object_scale * (1.-l.output[p_index]);
+
+                if(l.rescore){
+                    l.delta[p_index] = l.object_scale * (iou - l.output[p_index]);
+                }
+
+                l.delta[box_index+0] = l.coord_scale*(state.truth[tbox_index + 0] - l.output[box_index + 0]);
+                l.delta[box_index+1] = l.coord_scale*(state.truth[tbox_index + 1] - l.output[box_index + 1]);
+                l.delta[box_index+2] = l.coord_scale*(state.truth[tbox_index + 2] - l.output[box_index + 2]);
+                l.delta[box_index+3] = l.coord_scale*(state.truth[tbox_index + 3] - l.output[box_index + 3]);
+                if(l.sqrt){
+                    l.delta[box_index+2] = l.coord_scale*(sqrt(state.truth[tbox_index + 2]) - l.output[box_index + 2]);
+                    l.delta[box_index+3] = l.coord_scale*(sqrt(state.truth[tbox_index + 3]) - l.output[box_index + 3]);
+                }
+
+                *(l.cost) += pow(1-iou, 2);
+                avg_iou += iou;
+                ++count;
+            }
+            if(l.softmax){
+                gradient_array(l.output + index + locations*l.classes, locations*l.n*(1+l.coords), 
+                        LOGISTIC, l.delta + index + locations*l.classes);
             }
         }
-        printf("Avg IOU: %f\n", avg_iou/count);
+        printf("Detection Avg IOU: %f, Pos Cat: %f, All Cat: %f, Pos Obj: %f, Any Obj: %f, count: %d\n", avg_iou/count, avg_cat/count, avg_allcat/(count*l.classes), avg_obj/count, avg_anyobj/(l.batch*locations*l.n), count);
     }
 }
 
 void backward_detection_layer(const detection_layer l, network_state state)
 {
-    int locations = get_detection_layer_locations(l);
-    int i,j;
-    int in_i = 0;
-    int out_i = 0;
-    for(i = 0; i < l.batch*locations; ++i){
-        float scale = 1;
-        float latent_delta = 0;
-        if(l.joint) scale = state.input[in_i++];
-        else if (l.objectness)   state.delta[in_i++] += -l.delta[out_i++];
-        else if (l.background) state.delta[in_i++] += scale*l.delta[out_i++];
-        for(j = 0; j < l.classes; ++j){
-            latent_delta += state.input[in_i]*l.delta[out_i];
-            state.delta[in_i++] += scale*l.delta[out_i++];
-        }
-
-        if (l.objectness) {
-
-        }else if (l.background) gradient_array(l.output + out_i, l.coords, LOGISTIC, l.delta + out_i);
-        for (j = 0; j < l.coords; ++j){
-            state.delta[in_i++] += l.delta[out_i++];
-        }
-        if(l.joint) state.delta[in_i-l.coords-l.classes-l.joint] += latent_delta;
-    }
+    axpy_cpu(l.batch*l.inputs, 1, l.delta, 1, state.delta, 1);
 }
 
 #ifdef GPU
 
 void forward_detection_layer_gpu(const detection_layer l, network_state state)
 {
-    int outputs = get_detection_layer_output_size(l);
+    if(!state.train){
+        copy_ongpu(l.batch*l.inputs, state.input, 1, l.output_gpu, 1);
+        return;
+    }
+
     float *in_cpu = calloc(l.batch*l.inputs, sizeof(float));
     float *truth_cpu = 0;
     if(state.truth){
-        truth_cpu = calloc(l.batch*outputs, sizeof(float));
-        cuda_pull_array(state.truth, truth_cpu, l.batch*outputs);
+        int num_truth = l.batch*l.side*l.side*(1+l.coords+l.classes);
+        truth_cpu = calloc(num_truth, sizeof(float));
+        cuda_pull_array(state.truth, truth_cpu, num_truth);
     }
     cuda_pull_array(state.input, in_cpu, l.batch*l.inputs);
     network_state cpu_state;
@@ -176,38 +206,16 @@ void forward_detection_layer_gpu(const detection_layer l, network_state state)
     cpu_state.truth = truth_cpu;
     cpu_state.input = in_cpu;
     forward_detection_layer(l, cpu_state);
-    cuda_push_array(l.output_gpu, l.output, l.batch*outputs);
-    cuda_push_array(l.delta_gpu, l.delta, l.batch*outputs);
+    cuda_push_array(l.output_gpu, l.output, l.batch*l.outputs);
+    cuda_push_array(l.delta_gpu, l.delta, l.batch*l.inputs);
     free(cpu_state.input);
     if(cpu_state.truth) free(cpu_state.truth);
 }
 
 void backward_detection_layer_gpu(detection_layer l, network_state state)
 {
-    int outputs = get_detection_layer_output_size(l);
-
-    float *in_cpu    = calloc(l.batch*l.inputs, sizeof(float));
-    float *delta_cpu = calloc(l.batch*l.inputs, sizeof(float));
-    float *truth_cpu = 0;
-    if(state.truth){
-        truth_cpu = calloc(l.batch*outputs, sizeof(float));
-        cuda_pull_array(state.truth, truth_cpu, l.batch*outputs);
-    }
-    network_state cpu_state;
-    cpu_state.train = state.train;
-    cpu_state.input = in_cpu;
-    cpu_state.truth = truth_cpu;
-    cpu_state.delta = delta_cpu;
-
-    cuda_pull_array(state.input, in_cpu,    l.batch*l.inputs);
-    cuda_pull_array(state.delta, delta_cpu, l.batch*l.inputs);
-    cuda_pull_array(l.delta_gpu, l.delta, l.batch*outputs);
-    backward_detection_layer(l, cpu_state);
-    cuda_push_array(state.delta, delta_cpu, l.batch*l.inputs);
-
-    if (truth_cpu) free(truth_cpu);
-    free(in_cpu);
-    free(delta_cpu);
+    axpy_ongpu(l.batch*l.inputs, 1, l.delta_gpu, 1, state.delta, 1);
+    //copy_ongpu(l.batch*l.inputs, l.delta_gpu, 1, state.delta, 1);
 }
 #endif
