--- src/region_layer.c
@@ -1,259 +1,0 @@
-#include "region_layer.h"
-#include "activations.h"
-#include "softmax_layer.h"
-#include "blas.h"
-#include "box.h"
-#include "cuda.h"
-#include "utils.h"
-#include <stdio.h>
-#include <assert.h>
-#include <string.h>
-#include <stdlib.h>
-
-region_layer make_region_layer(int batch, int inputs, int n, int side, int classes, int coords, int rescore)
-{
-    region_layer l = {0};
-    l.type = REGION;
-
-    l.n = n;
-    l.batch = batch;
-    l.inputs = inputs;
-    l.classes = classes;
-    l.coords = coords;
-    l.rescore = rescore;
-    l.side = side;
-    assert(side*side*((1 + l.coords)*l.n + l.classes) == inputs);
-    l.cost = calloc(1, sizeof(float));
-    l.outputs = l.inputs;
-    l.truths = l.side*l.side*(1+l.coords+l.classes);
-    l.output = calloc(batch*l.outputs, sizeof(float));
-    l.delta = calloc(batch*l.outputs, sizeof(float));
-#ifdef GPU
-    l.output_gpu = cuda_make_array(l.output, batch*l.outputs);
-    l.delta_gpu = cuda_make_array(l.delta, batch*l.outputs);
-#endif
-
-    fprintf(stderr, "Region Layer\n");
-    srand(0);
-
-    return l;
-}
-
-void forward_region_layer(const region_layer l, network_state state)
-{
-    int locations = l.side*l.side;
-    int i,j;
-    memcpy(l.output, state.input, l.outputs*l.batch*sizeof(float));
-    int b;
-    if (l.softmax){
-        for(b = 0; b < l.batch; ++b){
-            int index = b*l.inputs;
-            for (i = 0; i < locations; ++i) {
-                int offset = i*l.classes;
-                softmax_array(l.output + index + offset, l.classes,
-                        l.output + index + offset);
-            }
-            int offset = locations*l.classes;
-            activate_array(l.output + index + offset, locations*l.n*(1+l.coords), LOGISTIC);
-        }
-    }
-    if (l.object_logistic) {
-        for(b = 0; b < l.batch; ++b){
-            int index = b*l.inputs;
-            int p_index = index + locations*l.classes;
-            activate_array(l.output + p_index, locations*l.n, LOGISTIC);
-        }
-    }
-
-    if (l.coord_logistic) {
-        for(b = 0; b < l.batch; ++b){
-            int index = b*l.inputs;
-            int coord_index = index + locations*(l.classes + l.n);
-            activate_array(l.output + coord_index, locations*l.n*l.coords, LOGISTIC);
-        }
-    }
-
-    if (l.class_logistic) {
-        for(b = 0; b < l.batch; ++b){
-            int class_index = b*l.inputs;
-            activate_array(l.output + class_index, locations*l.classes, LOGISTIC);
-        }
-    }
-
-    if(state.train){
-        float avg_iou = 0;
-        float avg_cat = 0;
-        float avg_allcat = 0;
-        float avg_obj = 0;
-        float avg_anyobj = 0;
-        int count = 0;
-        *(l.cost) = 0;
-        int size = l.inputs * l.batch;
-        memset(l.delta, 0, size * sizeof(float));
-        for (b = 0; b < l.batch; ++b){
-            int index = b*l.inputs;
-            for (i = 0; i < locations; ++i) {
-                int truth_index = (b*locations + i)*(1+l.coords+l.classes);
-                int is_obj = state.truth[truth_index];
-                for (j = 0; j < l.n; ++j) {
-                    int p_index = index + locations*l.classes + i*l.n + j;
-                    l.delta[p_index] = l.noobject_scale*(0 - l.output[p_index]);
-                    *(l.cost) += l.noobject_scale*pow(l.output[p_index], 2);
-                    avg_anyobj += l.output[p_index];
-                }
-
-                int best_index = -1;
-                float best_iou = 0;
-                float best_rmse = 20;
-
-                if (!is_obj){
-                    continue;
-                }
-
-                int class_index = index + i*l.classes;
-                for(j = 0; j < l.classes; ++j) {
-                    l.delta[class_index+j] = l.class_scale * (state.truth[truth_index+1+j] - l.output[class_index+j]);
-                    *(l.cost) += l.class_scale * pow(state.truth[truth_index+1+j] - l.output[class_index+j], 2);
-                    if(state.truth[truth_index + 1 + j]) avg_cat += l.output[class_index+j];
-                    avg_allcat += l.output[class_index+j];
-                }
-
-                box truth = float_to_box(state.truth + truth_index + 1 + l.classes);
-                truth.x /= l.side;
-                truth.y /= l.side;
-
-                for(j = 0; j < l.n; ++j){
-                    int box_index = index + locations*(l.classes + l.n) + (i*l.n + j) * l.coords;
-                    box out = float_to_box(l.output + box_index);
-                    out.x /= l.side;
-                    out.y /= l.side;
-
-                    if (l.sqrt){
-                        out.w = out.w*out.w;
-                        out.h = out.h*out.h;
-                    }
-
-                    float iou  = box_iou(out, truth);
-                    //iou = 0;
-                    float rmse = box_rmse(out, truth);
-                    if(best_iou > 0 || iou > 0){
-                        if(iou > best_iou){
-                            best_iou = iou;
-                            best_index = j;
-                        }
-                    }else{
-                        if(rmse < best_rmse){
-                            best_rmse = rmse;
-                            best_index = j;
-                        }
-                    }
-                }
-
-                if(l.forced){
-                    if(truth.w*truth.h < .1){
-                        best_index = 1;
-                    }else{
-                        best_index = 0;
-                    }
-                }
-
-                int box_index = index + locations*(l.classes + l.n) + (i*l.n + best_index) * l.coords;
-                int tbox_index = truth_index + 1 + l.classes;
-
-                box out = float_to_box(l.output + box_index);
-                out.x /= l.side;
-                out.y /= l.side;
-                if (l.sqrt) {
-                    out.w = out.w*out.w;
-                    out.h = out.h*out.h;
-                }
-                float iou  = box_iou(out, truth);
-
-                //printf("%d", best_index);
-                int p_index = index + locations*l.classes + i*l.n + best_index;
-                *(l.cost) -= l.noobject_scale * pow(l.output[p_index], 2);
-                *(l.cost) += l.object_scale * pow(1-l.output[p_index], 2);
-                avg_obj += l.output[p_index];
-                l.delta[p_index] = l.object_scale * (1.-l.output[p_index]);
-
-                if(l.rescore){
-                    l.delta[p_index] = l.object_scale * (iou - l.output[p_index]);
-                }
-
-                l.delta[box_index+0] = l.coord_scale*(state.truth[tbox_index + 0] - l.output[box_index + 0]);
-                l.delta[box_index+1] = l.coord_scale*(state.truth[tbox_index + 1] - l.output[box_index + 1]);
-                l.delta[box_index+2] = l.coord_scale*(state.truth[tbox_index + 2] - l.output[box_index + 2]);
-                l.delta[box_index+3] = l.coord_scale*(state.truth[tbox_index + 3] - l.output[box_index + 3]);
-                if(l.sqrt){
-                    l.delta[box_index+2] = l.coord_scale*(sqrt(state.truth[tbox_index + 2]) - l.output[box_index + 2]);
-                    l.delta[box_index+3] = l.coord_scale*(sqrt(state.truth[tbox_index + 3]) - l.output[box_index + 3]);
-                }
-
-                *(l.cost) += pow(1-iou, 2);
-                avg_iou += iou;
-                ++count;
-            }
-            if(l.softmax){
-                gradient_array(l.output + index + locations*l.classes, locations*l.n*(1+l.coords), 
-                        LOGISTIC, l.delta + index + locations*l.classes);
-            }
-            if (l.object_logistic) {
-                int p_index = index + locations*l.classes;
-                gradient_array(l.output + p_index, locations*l.n, LOGISTIC, l.delta + p_index);
-            }
-
-            if (l.class_logistic) {
-                int class_index = index;
-                gradient_array(l.output + class_index, locations*l.classes, LOGISTIC, l.delta + class_index);
-            }
-
-            if (l.coord_logistic) {
-                    int coord_index = index + locations*(l.classes + l.n);
-                gradient_array(l.output + coord_index, locations*l.n*l.coords, LOGISTIC, l.delta + coord_index);
-            }
-            //printf("\n");
-        }
-        printf("Region Avg IOU: %f, Pos Cat: %f, All Cat: %f, Pos Obj: %f, Any Obj: %f, count: %d\n", avg_iou/count, avg_cat/count, avg_allcat/(count*l.classes), avg_obj/count, avg_anyobj/(l.batch*locations*l.n), count);
-    }
-}
-
-void backward_region_layer(const region_layer l, network_state state)
-{
-    axpy_cpu(l.batch*l.inputs, 1, l.delta, 1, state.delta, 1);
-}
-
-#ifdef GPU
-
-void forward_region_layer_gpu(const region_layer l, network_state state)
-{
-    if(!state.train){
-        copy_ongpu(l.batch*l.inputs, state.input, 1, l.output_gpu, 1);
-        return;
-    }
-
-    float *in_cpu = calloc(l.batch*l.inputs, sizeof(float));
-    float *truth_cpu = 0;
-    if(state.truth){
-        int num_truth = l.batch*l.side*l.side*(1+l.coords+l.classes);
-        truth_cpu = calloc(num_truth, sizeof(float));
-        cuda_pull_array(state.truth, truth_cpu, num_truth);
-    }
-    cuda_pull_array(state.input, in_cpu, l.batch*l.inputs);
-    network_state cpu_state;
-    cpu_state.train = state.train;
-    cpu_state.truth = truth_cpu;
-    cpu_state.input = in_cpu;
-    forward_region_layer(l, cpu_state);
-    cuda_push_array(l.output_gpu, l.output, l.batch*l.outputs);
-    cuda_push_array(l.delta_gpu, l.delta, l.batch*l.inputs);
-    free(cpu_state.input);
-    if(cpu_state.truth) free(cpu_state.truth);
-}
-
-void backward_region_layer_gpu(region_layer l, network_state state)
-{
-    axpy_ongpu(l.batch*l.inputs, 1, l.delta_gpu, 1, state.delta, 1);
-    //copy_ongpu(l.batch*l.inputs, l.delta_gpu, 1, state.delta, 1);
-}
-#endif
-
